requirements.txt
# Microsoft Authentication and Graph API
msal>=1.26.0
requests>=2.31.0

# Document processing
python-docx>=1.1.0
PyPDF2>=3.0.1
pdfplumber>=0.10.3

# Environment and configuration
python-dotenv>=1.0.0

# Utilities
tqdm>=4.66.1  # Progress bars for downloads

# Optional but recommended
python-dateutil>=2.8.2  # For date parsing
